Checking for existence of mu-plugins folder before iterating

+++ src/psycle/Composer/WordPressCorePlugin.php
@@ -53,20 +53,20 @@ class WordPressCorePlugin implements PluginInterface, EventSubscriberInterface {
 				}
 			}
 		}
-		foreach ( new \DirectoryIterator($muInstallerPath) AS $directoryNode) {
-			if(!$directoryNode->isDot() && $directoryNode->isDir()) {
-				$muPluginDirectory = $directoryNode->getPathname();
-				$muBootstrapFile = dirname($muPluginDirectory) . DIRECTORY_SEPARATOR .'_' . basename($muPluginDirectory) . '.php';
+		if(file_exists($muInstallerPath)) {
+			foreach ( new \DirectoryIterator($muInstallerPath) AS $directoryNode) {
+				if(!$directoryNode->isDot() && $directoryNode->isDir()) {
+					$muPluginDirectory = $directoryNode->getPathname();
+					$muBootstrapFile = dirname($muPluginDirectory) . DIRECTORY_SEPARATOR .'_' . basename($muPluginDirectory) . '.php';
 
-				foreach (new \DirectoryIterator($muPluginDirectory) AS $muBootstrapableFile) {
-					if($muBootstrapableFile->isFile() && preg_match( '@\.php$@', $muBootstrapableFile )) {
-						$relativeFilePath = '__DIR__ . "/' . $directoryNode->getFilename() . '/' . $muBootstrapableFile->getFilename() . '"';
-						file_put_contents($muBootstrapFile, '<?php if(file_exists('.$relativeFilePath.')) require_once('.$relativeFilePath.');');
+					foreach (new \DirectoryIterator($muPluginDirectory) AS $muBootstrapableFile) {
+						if($muBootstrapableFile->isFile() && preg_match( '@\.php$@', $muBootstrapableFile )) {
+							$relativeFilePath = '__DIR__ . "/' . $directoryNode->getFilename() . '/' . $muBootstrapableFile->getFilename() . '"';
+							file_put_contents($muBootstrapFile, '<?php if(file_exists('.$relativeFilePath.')) require_once('.$relativeFilePath.');');
+						}
 					}
 				}
 			}
 		}
-		
-		
 	}
 }
